requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
SQLAlchemy==2.0.21
Werkzeug==2.3.7
openpyxl==3.1.2
python-dotenv==1.0.0
pytest==7.4.2
pytest-cov==4.1.0
gunicorn==21.2.0
Pillow==10.0.0
